refactor(types): move all types to types directory
