+++ src/Model.ts
@@ -2,6 +2,15 @@ import type { AxiosPromise } from 'axios'
 import { AxiosInstance, AxiosRequestConfig } from 'axios'
 
 import Builder from './Builder'
+import type {
+  ModelData,
+  RCollection,
+  RModel,
+  TCollection,
+  TModel,
+  WCollection,
+  WModel
+} from './types'
 import { getProp, setProp } from './utils'
 
 type Constructor<T extends Model<boolean, boolean>> = new (
@@ -11,45 +20,6 @@ type Constructor<T extends Model<boolean, boolean>> = new (
 type ThisClass<InstanceType extends Model<boolean, boolean>> = {
   new (...args: unknown[]): InstanceType
 }
-
-type ModelKeys = Omit<Model, 'delete' | 'save' | 'attach' | 'sync' | 'for'>
-
-type DerivedClass<T extends Model<boolean, boolean>> = Omit<T, keyof ModelKeys>
-
-export type ModelData<T extends Model<boolean, boolean>> = Required<
-  DerivedClass<T>
->
-
-type WModel<T extends Model<boolean, boolean>> = {
-  data: ModelData<T>
-}
-
-type TModel<T extends Model<boolean, boolean>> = ModelData<T> | WModel<T>
-
-type WCollection<T extends Model<boolean, boolean>> = {
-  data: TModel<T>[]
-}
-
-type TCollection<T extends Model<boolean, boolean>> =
-  | WCollection<T>
-  | TModel<T>[]
-
-export type RModel<
-  T extends Model<boolean, boolean>,
-  isWrappedModel
-> = isWrappedModel extends true ? WModel<T> : ModelData<T>
-
-type WRCollection<T extends Model<boolean, boolean>, isWrappedModel> = {
-  data: RModel<T, isWrappedModel>[]
-}
-
-export type RCollection<
-  T extends Model<boolean, boolean>,
-  isWrappedCollection,
-  isWrappedModel
-> = isWrappedCollection extends true
-  ? WRCollection<T, isWrappedModel>
-  : RModel<T, isWrappedModel>[]
 
 function hasProperty<T extends Model<boolean, boolean>>(
   obj: T,
